Fixed an Unhandled Error in badges.go
There was an unhandled error in badges.go. If you pass an invalid
branch to the badge handler (http://beta.drone.io/github.com/drone/drone/
status.svg?branch=this) the browser is told to redirect to a blank string.
This is unintended behavior. Now, a 404 response is sent instead.